+++ src/components/data/properties/material/sections/MatPropDisplacement.vue
@@ -1,0 +1,21 @@
+<template>
+	<MaterialInputFields :fields-list="fields" />
+</template>
+
+<script lang="ts" setup>
+import type { MaterialInputField, MeshMaterials } from '../utils/types'
+
+const fields: MaterialInputField<MeshMaterials>[] = [
+	{ type: 'map', label: 'Map', prop: 'displacementMap' },
+	{
+		type: 'number',
+		label: 'Scale',
+		prop: 'displacementScale'
+	},
+	{
+		type: 'number',
+		label: 'Bias',
+		prop: 'displacementBias'
+	}
+]
+</script>